requirements.txt
# Topology of Truth: Dependencies

# Core ML
#torch>=2.0.0
#transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0

# TDA (Topological Data Analysis)
ripser>=0.6.4
persim>=0.3.1

# Dimensionality reduction & visualization
umap-learn>=0.5.5
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Statistics
scipy>=1.12.0

# Data handling
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=15.0.0
pyyaml>=6.0.0

# API calls
requests>=2.31.0

# Progress bars
tqdm>=4.66.0
